move reachability tests out of frontend
